backend/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
python-dotenv==0.19.0
PyJWT==2.8.0
Werkzeug==3.0.1
bson==0.5.10
Flask-Mail==0.9.1
email-validator==2.1.0
requests==2.26.0
agora-token-builder==1.0.0
python-jose==3.3.0 